+++ lueur/platform/k8s/gateway.py
@@ -205,3 +205,65 @@ def expand_links(d: Discovery, serialized: dict[str, Any]) -> None:
                     pointer=str(urlmap.pointer()),
                 ),
             )
+
+    for gw_name in iter_resource(
+        serialized,
+        "$.resources[?@.meta.kind=='httproute' && @.meta.platform=='k8s'].struct.spec.parentRefs[?@.kind=='Gateway'].name",  # noqa E501
+    ):
+        r_id = gw_name.parent.parent.parent.parent.obj["id"]  # type: ignore
+        name = gw_name.value
+
+        p = f"$.resources[?@.meta.kind=='gateway' && @.meta.platform=='k8s' && @.meta.name=='{name}']"  # noqa E501
+        for gw in iter_resource(serialized, p):
+            add_link(
+                d,
+                r_id,
+                Link(
+                    direction="in",
+                    kind="gateway",
+                    path=gw.path,
+                    pointer=str(gw.pointer()),
+                ),
+            )
+
+    for gw_name in iter_resource(
+        serialized,
+        "$.resources[?@.meta.kind=='gateway' && @.meta.platform=='k8s'].meta.name",  # noqa E501
+    ):
+        r_id = gw_name.parent.parent.obj["id"]  # type: ignore
+        name = gw_name.value
+
+        p = f"$.resources[?@.meta.kind=='httproute' && @.meta.platform=='k8s'].struct.spec.parentRefs[?@.kind=='Gateway' && @.name=='{gw_name}']"  # noqa E501
+        for httproute in iter_resource(serialized, p):
+            add_link(
+                d,
+                r_id,
+                Link(
+                    direction="out",
+                    kind="httproute",
+                    path=httproute.path,
+                    pointer=str(httproute.pointer()),
+                ),
+            )
+
+    for service_refs in iter_resource(
+        serialized,
+        "$.resources[?@.meta.kind=='httproute' && @.meta.platform=='k8s'].struct.spec.rules.*.backendRefs[?@.kind=='Service']",  # noqa E501
+    ):
+        r_id = service_refs.parent.parent.parent.parent.parent.obj["id"]  # type: ignore
+        svc_ref = service_refs.obj
+        svc_name = svc_ref["name"]
+        svc_ns = svc_ref["namespace"]
+
+        p = f"$.resources[?@.meta.kind=='service' && @.meta.platform=='k8s' && @.meta.ns=='{svc_ns}' && @.meta.name=='{svc_name}']"  # noqa E501
+        for svc in iter_resource(serialized, p):
+            add_link(
+                d,
+                r_id,
+                Link(
+                    direction="out",
+                    kind="service",
+                    path=svc.path,
+                    pointer=str(svc.pointer()),
+                ),
+            )